prompt: default prompt for KG builder
Previously, due to the fact we assumed "name" is special for
a extracted node to actually act as its vertexID, and, there
in NebulaGraph the id is actually composite 4 fields(src,dst,type,rank),
thus there is no "name" fields in extracted JSON.

In such case, the "name" field in schema of edge type will not
be handled properly in some of LLMs.

Also, previously, the JSON format was not exactly an example, but
a half example(nodes, edges) half json type schema(xxx:string, yyy:object).

This is actually not clear enough, which will cost mind-power of LLMs.

Now it's changed to a real example.